requirements.txt
Flask==2.1.2
requests==2.27.1
python-dotenv==0.20.0
asyncio==3.4.3